fix(update): address code-review findings on the channel-aware update check
- doctor: inject DaemonConfig for the channel instead of hand-reading
  netclaw.json, so a non-string Daemon.UpdateChannel no longer crashes the
  whole `netclaw doctor` run (findings #1, #4).
- update-check: drop the low-value 1h TTL and keep a plain last-result store
  for the daemon /status API (finding #3). Make `channel` required (no default)
  on EvaluateManifest + CheckForUpdateAsync so a forgotten arg can't silently
  fall back to stable (finding #9).
- SemVer: parse numeric prerelease identifiers as long, matching the bash
  generator's unbounded ints (finding #6).
- /status: report FullVersion in the no-check-yet branch for consistency with
  the post-check branch (finding #8).
- dotted prerelease convention (beta.N): update docs/examples + widen the
  property-test generators + add example cases; the release version gate now
  rejects mixed identifiers like `beta1` so a non-dotted tag can't ship and
  silently mis-order the channel (finding #2).
- conformance: extract the generator's precedence key to
  feeds/scripts/semver_key.py and assert BOTH it and the C# SemVer comparator
  order one shared fixture (feeds/scripts/semver-order.txt) — via a C# test and
  a CI check — so the two implementations can't drift (finding #7).

Also updates the release-channels OpenSpec change (dotted-tag requirement plus
risk/decision notes).

+++ src/Netclaw.Configuration/Feeds/SemVer.cs
@@ -115,7 +115,7 @@ public static class SemVer
         if (c != 0) return c;
 
         // A version with no prerelease outranks one that has a prerelease with the
-        // same core (1.0.0 > 1.0.0-beta1).
+        // same core (1.0.0 > 1.0.0-beta.1).
         var aPre = a.Pre.Length > 0;
         var bPre = b.Pre.Length > 0;
         if (!aPre && !bPre) return 0;
@@ -137,8 +137,11 @@ public static class SemVer
 
     private static int ComparePreIdentifier(string a, string b)
     {
-        var aNumeric = int.TryParse(a, out var ai);
-        var bNumeric = int.TryParse(b, out var bi);
+        // Parse as long (not int): numeric identifiers can be large (timestamps, build
+        // counters), and the bash generator compares them with Python's unbounded int —
+        // long covers any realistic value and keeps the two implementations in agreement.
+        var aNumeric = long.TryParse(a, out var ai);
+        var bNumeric = long.TryParse(b, out var bi);
 
         if (aNumeric && bNumeric) return ai.CompareTo(bi);
         // Numeric identifiers always have lower precedence than alphanumeric ones.
